pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kicomav"
version = "0.40"
description = "KICOM Anti-Virus II - Open source antivirus engine for Python"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Kei Choi", email = "hanul93@gmail.com"}
]
keywords = ["antivirus", "malware", "security", "scanner", "virus"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "yara-python>=4.3.0",
    "py7zr>=0.20.0",
    "rarfile>=4.0",
    "pycabfile>=0.1.0",
]

[project.optional-dependencies]
full = [
    "pylzma>=0.5.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "coverage>=7.0.0",
]

[project.scripts]
k2 = "kicomav.k2:main"

[project.urls]
Homepage = "https://github.com/hanul93/kicomav"
Repository = "https://github.com/hanul93/kicomav"
Documentation = "https://github.com/hanul93/kicomav#readme"
"Bug Tracker" = "https://github.com/hanul93/kicomav/issues"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["kicomav*"]

[tool.setuptools.package-data]
"kicomav.plugins" = ["*.py", "kicom.lst"]
"kicomav.kavcore" = ["*.py"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312", "py313"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
